+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # vdiffr 0.2.2.9000
 
+* Use `last_collection_error()` to print a testthat error that
+  occurred while collecting the test cases.
+
 * Added global options to disable vdiffr. Set `vdiffr_skip` to `TRUE`
   to skip all vdiffr expectations.